+++ Sources/BraintreePayPal/BTPayPalClient.swift
@@ -213,6 +213,7 @@ import BraintreeDataCollector
 
         apiClient.sendAnalyticsEvent(
             BTPayPalAnalytics.handleReturnStarted,
+            applicationState: UIApplication.shared.applicationStateString,
             appSwitchURL: url,
             contextID: contextID,
             contextType: contextType,
@@ -296,6 +297,7 @@ import BraintreeDataCollector
         if success {
             apiClient.sendAnalyticsEvent(
                 BTPayPalAnalytics.appSwitchSucceeded,
+                applicationState: UIApplication.shared.applicationStateString,
                 appSwitchURL: url,
                 contextID: contextID,
                 contextType: contextType,
@@ -312,6 +314,7 @@ import BraintreeDataCollector
 
             apiClient.sendAnalyticsEvent(
                 BTPayPalAnalytics.appSwitchFailed,
+                applicationState: UIApplication.shared.applicationStateString,
                 appSwitchURL: url,
                 contextID: contextID,
                 contextType: contextType,
@@ -358,6 +361,7 @@ import BraintreeDataCollector
 
         apiClient.sendAnalyticsEvent(
             BTPayPalAnalytics.tokenizeStarted,
+            applicationState: UIApplication.shared.applicationStateString,
             contextType: contextType,
             didEnablePayPalAppSwitch: payPalRequest?.enablePayPalAppSwitch,
             isVaultRequest: isVaultRequest,
@@ -451,6 +455,7 @@ import BraintreeDataCollector
         guard !hasOpenedURL else {
             apiClient.sendAnalyticsEvent(
                 BTPayPalAnalytics.tokenizeDuplicateRequest,
+                applicationState: UIApplication.shared.applicationStateString,
                 appSwitchURL: payPalAppRedirectURL,
                 contextID: contextID,
                 contextType: contextType,
@@ -468,6 +473,7 @@ import BraintreeDataCollector
 
         apiClient.sendAnalyticsEvent(
             BTPayPalAnalytics.appSwitchStarted,
+            applicationState: UIApplication.shared.applicationStateString,
             appSwitchURL: payPalAppRedirectURL,
             contextID: contextID,
             contextType: contextType,
@@ -505,6 +511,7 @@ import BraintreeDataCollector
     ) {
         apiClient.sendAnalyticsEvent(
             BTPayPalAnalytics.browserPresentationStarted,
+            applicationState: UIApplication.shared.applicationStateString,
             appSwitchURL: appSwitchURL,
             contextID: contextID,
             contextType: contextType,
@@ -554,6 +561,7 @@ import BraintreeDataCollector
             if didAppear {
                 apiClient.sendAnalyticsEvent(
                     BTPayPalAnalytics.browserPresentationSucceeded,
+                    applicationState: UIApplication.shared.applicationStateString,
                     appSwitchURL: appSwitchURL,
                     contextID: contextID,
                     contextType: contextType,
@@ -566,6 +574,7 @@ import BraintreeDataCollector
             } else {
                 apiClient.sendAnalyticsEvent(
                     BTPayPalAnalytics.browserPresentationFailed,
+                    applicationState: UIApplication.shared.applicationStateString,
                     appSwitchURL: appSwitchURL,
                     contextID: contextID,
                     contextType: contextType,
@@ -582,6 +591,7 @@ import BraintreeDataCollector
                 // User tapped system cancel button on permission alert
                 apiClient.sendAnalyticsEvent(
                     BTPayPalAnalytics.browserLoginAlertCanceled,
+                    applicationState: UIApplication.shared.applicationStateString,
                     appSwitchURL: appSwitchURL,
                     contextID: contextID,
                     contextType: contextType,
@@ -600,6 +610,7 @@ import BraintreeDataCollector
             
             apiClient.sendAnalyticsEvent(
                 BTPayPalAnalytics.tokenizeDuplicateRequest,
+                applicationState: UIApplication.shared.applicationStateString,
                 appSwitchURL: appSwitchURL,
                 contextID: contextID,
                 contextType: contextType,
@@ -635,6 +646,7 @@ import BraintreeDataCollector
     ) {
         apiClient.sendAnalyticsEvent(
             BTPayPalAnalytics.tokenizeSucceeded,
+            applicationState: UIApplication.shared.applicationStateString,
             contextID: contextID,
             contextType: contextType,
             correlationID: contextID.flatMap { clientMetadataIDs[$0] },
@@ -649,6 +661,7 @@ import BraintreeDataCollector
     private func notifyFailure(with error: Error, completion: @escaping (BTPayPalAccountNonce?, Error?) -> Void) {
         apiClient.sendAnalyticsEvent(
             BTPayPalAnalytics.tokenizeFailed,
+            applicationState: UIApplication.shared.applicationStateString,
             contextID: contextID,
             contextType: contextType,
             correlationID: contextID.flatMap { clientMetadataIDs[$0] },