+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Braintree iOS SDK Release Notes
 
+## 6.38.0 (2025-09-09)
+* BraintreeShopperInsights (BETA)
+  * Add `createCustomerSession(request:)` to `BTShopperInsightsClientV2`
+  * Add `updateCustomerSession(request:sessionID:)` to `BTShopperInsightsClientV2`
+  * Add `generateCustomerRecommendations(request:sessionID:)` to `BTShopperInsightsClientV2`
+  * Add `BTPageType.paymentMethodSelection` and `BTPageType.paymentMethodAddition`
+* BraintreePayPal
+  * Send `application_state` to FPTI for tracking UIApplication state
+
 ## 6.37.0 (2025-08-26)
 * BraintreePayPal
   * Fix an issue where `BTPayPalRequest` was sending `phone_number` instead of `payer_phone`